+++ MatrixKit/Controllers/MXKRoomViewController.m
@@ -51,6 +51,7 @@
 #import "NSBundle+MatrixKit.h"
 #import "UIScrollView+MatrixKit.h"
 #import "MXKSlashCommands.h"
+#import "MXKSwiftHeader.h"
 
 @interface MXKRoomViewController ()
 {
@@ -143,9 +144,9 @@
 @property (nonatomic, weak) id mxSessionWillLeaveRoomNotificationObserver;
 
 /**
- Observe UIApplicationWillEnterForegroundNotification to refresh bubbles when app leaves the background state.
+ Observe UIApplicationDidBecomeActiveNotification to refresh bubbles when app leaves the background state.
  */
-@property (nonatomic, weak) id uiApplicationWillEnterForegroundNotificationObserver;
+@property (nonatomic, weak) id uiApplicationDidBecomeActiveNotificationObserver;
 
 /**
  Observe UIMenuControllerDidHideMenuNotification to cancel text selection
@@ -269,9 +270,9 @@
     // Finalize table view configuration
     [self configureBubblesTableView];
     
-    // Observe UIApplicationWillEnterForegroundNotification to refresh bubbles when app leaves the background state.
+    // Observe UIApplicationDidBecomeActiveNotification to refresh bubbles when app leaves the background state.
     MXWeakify(self);
-    _uiApplicationWillEnterForegroundNotificationObserver = [[NSNotificationCenter defaultCenter] addObserverForName:UIApplicationWillEnterForegroundNotification object:nil queue:[NSOperationQueue mainQueue] usingBlock:^(NSNotification *notif) {
+    _uiApplicationDidBecomeActiveNotificationObserver = [[NSNotificationCenter defaultCenter] addObserverForName:UIApplicationDidBecomeActiveNotification object:nil queue:[NSOperationQueue mainQueue] usingBlock:^(NSNotification *notif) {
         
         MXStrongifyAndReturnIfNil(self);
         if (self->roomDataSource.state == MXKDataSourceStateReady && [self->roomDataSource tableView:self->_bubblesTableView numberOfRowsInSection:0])
@@ -378,9 +379,9 @@
         [[NSNotificationCenter defaultCenter] removeObserver:_mxSessionWillLeaveRoomNotificationObserver];
     }
     
-    if (_uiApplicationWillEnterForegroundNotificationObserver)
+    if (_uiApplicationDidBecomeActiveNotificationObserver)
     {
-        [[NSNotificationCenter defaultCenter] removeObserver:_uiApplicationWillEnterForegroundNotificationObserver];
+        [[NSNotificationCenter defaultCenter] removeObserver:_uiApplicationDidBecomeActiveNotificationObserver];
     }
     
     if (_uiMenuControllerDidHideMenuNotificationObserver)
@@ -2858,7 +2859,7 @@
                                                                    
                                                                    if (textMessage)
                                                                    {
-                                                                       [[UIPasteboard generalPasteboard] setString:textMessage];
+                                                                       MXKPasteboardManager.shared.pasteboard.string = textMessage;
                                                                    }
                                                                    else
                                                                    {
@@ -3181,7 +3182,7 @@
 {
     if (selectedText)
     {
-        [[UIPasteboard generalPasteboard] setString:selectedText];
+        MXKPasteboardManager.shared.pasteboard.string = selectedText;
     }
     else
     {